+++ CHANGELOG.md
@@ -8,7 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- Add new api client for configmgr api calls
 - Add new resource aem_configmgr
+
+### Changed
+- Change SAML Api calls to use configmgr api
 
 ## 3.2.1 - 2019-06-05
 ### Fixed